\name{biodecrypt.plot}
\alias{biodecrypt.plot}
\title{
Plot biodecrypt and biodecrypt.cross results
}
\description{
Plots the results of \code{biodecrypt} and \code{biodecrypt.cross}. The function adds coloured occurrence symbols to an existing plot, optionally drawing taxon hulls and distinguishing known, attributed, non-attributed, misidentified and non-identified records.
}
\usage{
biodecrypt.plot(x, minsize = 0.3, pchid = 1, cexid = 0.1,
                square = 0.001,
                col = c("red", "darkgreen", "blue", "purple"),
                attributed = NULL, hull = TRUE, NUR = "black",
                fading = 50, ...)
}
\arguments{
  \item{x}{
An object returned by \code{biodecrypt} or \code{biodecrypt.cross}.
}
  \item{minsize}{
Minimum size of plotted symbols.
}
  \item{pchid}{
Plotting character used to mark a priori identified records when \code{attributed = "points"}.
}
  \item{cexid}{
Size of the points marking a priori identified records when \code{attributed = "points"}.
}
  \item{square}{
Size of the grid cell used to collapse nearby occurrence records into pies. If the value is smaller than the data resolution, records are effectively plotted separately.
}
  \item{col}{
Colours assigned to taxa. Colours are matched to the internal taxon order stored in the \code{biodecrypt} result. If fewer colours than taxa are provided, colours are recycled.
}
  \item{attributed}{
Method used to distinguish records attributed by \code{biodecrypt} from records identified a priori. With \code{attributed = "fade"}, attributed records are plotted with paler colours. With \code{attributed = "points"}, a point is added to mark a priori identified records. If \code{NULL}, \code{"fade"} is used.
}
  \item{hull}{
Logical. If \code{TRUE}, hulls are added to the plot.
}
  \item{NUR}{
Colour used for non-attributed unidentified records.
}
  \item{fading}{
Degree of fading used for attributed records when \code{attributed = "fade"}. Larger values make colours paler.
}
  \item{...}{
Additional arguments passed to the underlying plotting functions.
}
}
\details{
The function is designed to be used after a base plot has already been created, for example with \code{plot()} on a map or on the occurrence coordinates.

For \code{biodecrypt} objects, known records, attributed records and non-attributed unidentified records are plotted using pies generated by \code{recluster.plot.pie}. A priori identified records can be distinguished either by fading attributed records or by adding points on known records.

For \code{biodecrypt.cross} objects, misidentified records (MIR) are plotted in black and non-identified records (NIR) in white.

The current implementation supports non-consecutive numeric taxon identifiers and character taxon identifiers, provided that the input object was generated by the updated \code{biodecrypt} function. Internally, colours are matched using the numeric taxon codes stored in \code{x$table_internal}, while original taxon labels are retained in \code{x$table}.
}
\value{
The function is called for its side effect and adds graphical elements to the active plot.
}
\references{
Platania L. et al. Assigning occurrence data to cryptic taxa improves climatic niche assessments: biodecrypt, a new tool tested on European butterflies. Global Ecology and Biogeography (2020).
}
\author{
Leonardo Dapporto
}
\examples{
# See examples in biodecrypt and biodecrypt.cross.
}
\keyword{record attribution}
\keyword{spatial}
\keyword{plot}
