\name{dataisl}
\alias{dataisl}
\docType{data}
\title{
West Mediterranean island butterflies provided with the package recluster 
}
\description{
This dataset represent presence of butterfly species in 33 West-Mediterranean islands
}
\usage{
data(dataisl)
}
\details{
  A data frame with 33 observations (islands) on 122 binary variables (species).
}
\source{
Dapporto L., Ramazzotti M., Fattorini S., Vila R., Dennis R. 
"recluster: an unbiased clustering procedure to detect biogeographic patterns by beta-diversity turnover"
XXXX (2013), XXX:XXX.
\url{
www.unifi.it/scibio/bioinfo/recluster.html
}
}
\keyword{datasets}
