\name{datamod}
\alias{datamod}
\docType{data}
\title{
Virtual island faunas provided with the package recluster 
}
\description{
This dataset represent a series of virtual faunas in different sites
}
\usage{
data(datamod)
}
\details{
  A data frame with 9 observations (sites) on 31 binary variables (species).
}
\source{
Dapporto L., Ramazzotti M., Fattorini S., Vila R., Talavera G., Dennis R. 
"recluster: an unbiased clustering procedure for beta-diversity turnover"
XXXX (2013), XXX:XXX.
\url{
www.unifi.it/scibio/bioinfo/recluster.html
}
}
\keyword{datasets}
