\name{multiboot}
\alias{multiboot}
\docType{data}
\title{
A multiboot result from dataisl dataset.
}
\description{
This dataset represent an output for a multiscale bootstrap composed of 30 scales (x1-x30). 
}
\usage{
data(dataisl)
}
\details{
  A data frame with 29 nodes (rows) and 30 different scales of bootstrap(columns). NAs values represent collapsed nodes
}
\source{
Dapporto L., Ramazzotti M., Fattorini S., Vila R., Talavera G., Dennis R. 
"recluster: an unbiased clustering procedure for beta-diversity turnover"
XXXX (2013), XXX:XXX.
\url{
www.unifi.it/scibio/bioinfo/recluster.html
}
}
\author{
Leonardo Dapporto
}
\keyword{datasets}
