\name{recluster.check}
\alias{recluster.check}
\title{
An internal workaround for the new ape package
}
\description{
This function solve a label problem producing unreliable low values of bootstrap
}
\usage{
recluster.check(tree, tip)
}

\arguments{
  \item{tree}{
An object of class phylo
}
\item{tip}{
A tip name
}
}
\value{
An oject of the class phylo
}
\references{
Dapporto L., Ramazzotti M., Fattorini S., Vila R., Talavera G., Dennis R. 
"recluster: an unbiased clustering procedure for beta-diversity turnover"
XXXX (2013), XXX:XXX.
\url{
www.unifi.it/scibio/bioinfo/recluster.html
}
}
\author{
Leonardo Dapporto and Matteo Ramazzotti
}
\keyword{ internal }
