\name{recluster.phybetadiv}
\alias{recluster.phybetadiv}
\title{
An series of function for phylogenetic beta diversity
}
\description{
This series of functions serve to compute dissimilarity matrices for phylogenetic beta diversity
}
\usage{
.................
}

\arguments{
  \item{tree}{
An object of class phylo
}
\item{tip}{
A tip name
}
}
\value{
An oject of the class phylo
}
\references{
Leprieur F., Albouy C., De Bortoli J., Cowman P.F., Bellwood D.R., et al. Quantifying Phylogenetic Beta Diversity: Distinguishing between "True" Turnover of Lineages and Phylogenetic Diversity Gradients. PLoS ONE (2012), 7(8), e42760. doi:10.1371/journal.pone.0042760

Dapporto L., Ramazzotti M., Fattorini S., Vila R., Talavera G., Dennis R. 
"recluster: an unbiased clustering procedure for beta-diversity turnover"
XXXX (2013), XXX:XXX.
\url{
www.unifi.it/scibio/bioinfo/recluster.html
}
}
\author{
modified by Leonadro Dapporto from Leprieur et al., 2013
}
\keyword{ internal }
