\name{treemod}
\alias{treemod}
\docType{data}
\title{
Hypothetical phylogenetic tree for the virtual island faunas provided with the package recluster 
}
\description{
This phylogenetic tree has been created from the datamod dataset representing a series of virtual faunas in different sites
}
\usage{
data(treemod)
}
\details{
  A phylogenetic tree of 31 species taken from 9 sites.
}
\source{
Dapporto L., Ramazzotti M., Fattorini S., Vila R., Talavera G., Dennis R. 
"recluster: an unbiased clustering procedure for beta-diversity turnover"
XXXX (2013), XXX:XXX.
\url{
www.unifi.it/scibio/bioinfo/recluster.html
}
}
\author{
Gerard Talavera
}
\keyword{datasets}
