\name{biodecrypt.view}
\alias{biodecrypt.view}
\title{
A function to view the hulls and manually adjust alpha values before applying biodecrypt
}
\description{
biodecrypt.view is deprecated and no longer supported. Please use biodecrypt with plot=T to achieve the same result. This function will be removed in future versions.
}
