\name{dataisl}
\alias{dataisl}
\docType{data}
\title{
West Mediterranean island butterflies provided with the package recluster 
}
\description{
This dataset represents occurrence data of butterfly species in 30 West-Mediterranean islands
}
\usage{
data(dataisl)
}
\details{
  A data frame with 30 observations (islands) on 123 binary variables (species).
}
\references{
Dapporto L., Ramazzotti M., Fattorini S., Talavera G., Vila R., Dennis R. 
"recluster: an unbiased clustering procedure for beta-diversity turnover"
Ecography (2013), 36:1070-1075.
}
\author{
Leonardo Dapporto and Roger Vila
}
\keyword{datasets}
