\name{datamod}
\alias{datamod}
\docType{data}
\title{
Virtual island faunas provided with the package recluster 
}
\description{
This dataset represents a series of virtual faunas in different sites
}
\usage{
data(datamod)
}
\details{
  A data frame with 9 observations (sites) on 31 binary variables (species).
}
\references{
Dapporto L., Ramazzotti M., Fattorini S., Talavera G., Vila R., Dennis R. 
"recluster: an unbiased clustering procedure for beta-diversity turnover"
Ecography (2013), 36:1070-1075.
}
\keyword{datasets}
