\name{multiboot}
\alias{multiboot}
\docType{data}
\title{
A multiboot result obtained with the dataisl dataset.
}
\description{
This dataset represents an output for a multiscale bootstrap composed of 30 scales (x1-x30). 
}
\usage{
data(dataisl)
}
\details{
  A data frame with 29 nodes (rows) and 30 different scales of bootstrap(columns). NAs values represent collapsed nodes
}
\source{
Dapporto L., Ramazzotti M., Fattorini S., Talavera G., Vila R., Dennis R. 
"recluster: an unbiased clustering procedure for beta-diversity turnover"
Ecography (2013), 36:1070-1075.
}
\author{
Leonardo Dapporto
}
\keyword{datasets}
