\name{treebut}
\alias{treebut}
\docType{data}
\title{
Phylogenetic tree for the butterfly species included in dataisl dataset 
}
\description{
This phylogenetic tree has been created based on known phylogeny of butterflies at family and subfamily level and on COI sequences at genus and species level. Branch lenghts have been calculated by Graphen method
}
\usage{
data(treemod)
}
\details{
  A phylogenetic tree of butterfly species occurring on Western Mediterranean islands.
}
\references{
Dapporto L., Ramazzotti M., Fattorini S., Talavera G., Vila R., Dennis R. 
"recluster: an unbiased clustering procedure for beta-diversity turnover"
Ecography (2013), 36:1070-1075.
}
\author{
Gerard Talavera and Roger Vila
}
\keyword{datasets}
